feat!: add support for validation and error output (#71)
Adds support for shader semantic validation. This is in service of
ensuring that an embedded shader will actually load at runtime. Error
messages can also be formatted (either to stdout or to a string) in a
consistent way that matches the output of rustc and other build tools.
`CreateModuleError` has been marked as non-exhaustive to reduce breaking
changes in the future.